refactor(wiki): extract lib/fileview/ for read-mode rendering
